refactor: move group-node migration types out of core litegraph
Remove dead updateLink augmentation (no callers) and relocate the
group-node-only GroupNodeConfigEntry/GroupNodeWorkflowData types from
core LGraph.ts into groupNode.ts. The serialized extra.groupNodes field
remains reachable via the LGraphExtra Dictionary index signature, so
core no longer references any group-node type.

Amp-Thread-ID: https://ampcode.com/threads/T-019edc21-0650-748d-8ec7-50778c0f4441
